pyproject.toml
[project]
name = "image_processing"
version = "0.1.0"
description = "Image processing utilities with edge detection, sharpening, and filtering"
authors = [
    {name = "Image Processing Team"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "opencv-python>=4.8.0",
    "imageio>=2.31.1",
    "matplotlib>=3.7.1",
    "numpy>=1.24.3",
    "Pillow>=10.0.0",
    "ipython>=8.0.0",
    "scipy>=1.10.0",
    "bs4>=0.0.2",
    "PySimpleGUI>=5.0.8.2",
    "PyInstaller>=6.13.0",
    "customtkinter>=5.2.2",
    "playwright>=1.52.0",
    "watchdog>=6.0.0",
    "pika>=1.3.2",
]

[project.scripts]
image-processor = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pip]
no-dependencies = true
